Modify js files to remove classes
