--- module/constructor.js
@@ -1,8 +1,0 @@
-/* eslint-disable linebreak-style */
-export default class Book {
-  constructor(title, author, id) {
-    this.title = title;
-    this.author = author;
-    this.id = id;
-  }
-}
